+++ controllers/contacts-controller.js
@@ -1,10 +1,6 @@
 import Contact from "../models/contact.js";
 import ctrlWrapper from "../decorators/ctrlWrapper.js";
 import httpError from "../helpers/httpError.js";
-import fs from "fs/promises";
-import path from "path";
-
-const avatarPath = path.resolve("public", "avatars");
 
 const getAllContacts = async (req, res) => {
   const { _id: owner } = req.user;
@@ -32,11 +28,7 @@ const getContactsById = async (req, res) => {
 
 const addContact = async (req, res) => {
   const { _id: owner } = req.user;
-  const { path: oldPath, filename } = req.file;
-  const newPath = path.join(avatarPath, filename);
-  await fs.rename(oldPath, newPath);
-  const avatarURL = path.join("avatars", filename);
-  const result = await Contact.create({ ...req.body, avatarURL, owner });
+  const result = await Contact.create({ ...req.body, owner });
   res.status(201).json(result);
 };
