+++ middlewars/upload.js
@@ -1,6 +1,6 @@
 import multer from "multer";
 import path from "path";
-const destination = path.resolve("temp");
+const destination = path.resolve("tmp");
 
 const storage = multer.diskStorage({
   destination,
@@ -11,13 +11,8 @@ const storage = multer.diskStorage({
   },
 });
 
-const limits = {
-  fileSize: 1024 * 1024 * 5,
-};
-
 const upload = multer({
   storage,
-  limits,
 });
 
 export default upload;
