+++ models/contact.js
@@ -17,7 +17,6 @@ const contactSchema = new Schema(
       type: Boolean,
       default: false,
     },
-    avatarURL: String,
     owner: {
       type: Schema.Types.ObjectId,
       ref: "user",
