+++ package.json
@@ -16,6 +16,7 @@
     "dotenv": "^16.3.1",
     "express": "4.17.1",
     "gravatar": "^1.8.2",
+    "jimp": "^0.22.10",
     "joi": "^17.9.2",
     "jsonwebtoken": "^9.0.1",
     "mongoose": "^7.4.0",
