+++ routes/api/auth.js
@@ -3,6 +3,8 @@ import validateBody from "../../decorators/validateBody.js";
 import usersSchemes from "../../schemes/users-schemes.js";
 import authContoller from "../../controllers/auth-contoller.js";
 import authenticate from "../../middlewars/authenticate.js";
+import upload from "../../middlewars/upload.js";
+import ctrlWrapper from "../../decorators/ctrlWrapper.js";
 
 const authRouter = express.Router();
 
@@ -27,6 +29,13 @@ authRouter.patch(
   authenticate,
   validateBody(usersSchemes.userSubUpdateSchema),
   authContoller.subUpdate
+);
+
+authRouter.patch(
+  "/avatars",
+  authenticate,
+  upload.single("avatar"),
+  authContoller.avatarUpdate
 );
 
 export default authRouter;
